requirements.txt
pyserial
pywaggle